+++ wpi_jaco_wrapper/src/jaco_arm_trajectory_node.cpp
@@ -21,15 +21,25 @@ JacoArmTrajectoryController::JacoArmTrajectoryController(ros::NodeHandle nh, ros
 
   boost::recursive_mutex::scoped_lock lock(api_mutex);
 
+  //ROS_INFO("Trying to initialize JACO API...");
   InitAPI();
+  //ROS_INFO("Api initialized.");
   ros::Duration(1.0).sleep();
+  //ROS_INFO("Starting control API...");
   StartControlAPI();
+  //ROS_INFO("Control API started...");
   ros::Duration(3.0).sleep();
+  //ROS_INFO("Stopping control API...");
   StopControlAPI();
+  //ROS_INFO("Control API stopped.");
 
   // Initialize arm
+  //ROS_INFO("Homing arm...");
   MoveHome();
+  //ROS_INFO("Done.");
+  //ROS_INFO("Initializing fingers...");
   InitFingers();
+  //ROS_INFO("Done.");
   SetFrameType(0); //set end effector to move with respect to the fixed frame
 
   // Initialize joint names
@@ -720,7 +730,6 @@ void JacoArmTrajectoryController::execute_gripper(const control_msgs::GripperCom
 void JacoArmTrajectoryController::home_arm(const wpi_jaco_msgs::HomeArmGoalConstPtr &goal)
 {
   boost::recursive_mutex::scoped_lock lock(api_mutex);
-
   StopControlAPI();
   MoveHome();
   StartControlAPI();
@@ -835,13 +844,21 @@ void JacoArmTrajectoryController::angularCmdCallback(const wpi_jaco_msgs::Angula
   }
   else
   {
-    jacoPoint.Position.Type = ANGULAR_VELOCITY;
-    jacoPoint.Position.Actuators.Actuator1 = 0.0;
-    jacoPoint.Position.Actuators.Actuator2 = 0.0;
-    jacoPoint.Position.Actuators.Actuator3 = 0.0;
-    jacoPoint.Position.Actuators.Actuator4 = 0.0;
-    jacoPoint.Position.Actuators.Actuator5 = 0.0;
-    jacoPoint.Position.Actuators.Actuator6 = 0.0;
+    if (msg.position)
+    {
+      fingerPositionControl(msg.fingers[0], msg.fingers[1], msg.fingers[2]);
+      return;
+    }
+    else
+    {
+      jacoPoint.Position.Type = ANGULAR_VELOCITY;
+      jacoPoint.Position.Actuators.Actuator1 = 0.0;
+      jacoPoint.Position.Actuators.Actuator2 = 0.0;
+      jacoPoint.Position.Actuators.Actuator3 = 0.0;
+      jacoPoint.Position.Actuators.Actuator4 = 0.0;
+      jacoPoint.Position.Actuators.Actuator5 = 0.0;
+      jacoPoint.Position.Actuators.Actuator6 = 0.0;
+    }
   }
 
   //populate finger command
@@ -915,15 +932,23 @@ void JacoArmTrajectoryController::cartesianCmdCallback(const wpi_jaco_msgs::Cart
   }
   else
   {
-    jacoPoint.Position.Type = CARTESIAN_VELOCITY;
-    jacoPoint.Position.CartesianPosition.X = 0.0;
-    jacoPoint.Position.CartesianPosition.Y = 0.0;
-    jacoPoint.Position.CartesianPosition.Z = 0.0;
-    jacoPoint.Position.CartesianPosition.ThetaX = 0.0;
-    jacoPoint.Position.CartesianPosition.ThetaY = 0.0;
-    jacoPoint.Position.CartesianPosition.ThetaZ = 0.0;
+    if (msg.position)
+    {
+      fingerPositionControl(msg.fingers[0], msg.fingers[1], msg.fingers[2]);
+      return;
+    }
+    else
+    {
+      jacoPoint.Position.Type = ANGULAR_VELOCITY;
+      jacoPoint.Position.Actuators.Actuator1 = 0.0;
+      jacoPoint.Position.Actuators.Actuator2 = 0.0;
+      jacoPoint.Position.Actuators.Actuator3 = 0.0;
+      jacoPoint.Position.Actuators.Actuator4 = 0.0;
+      jacoPoint.Position.Actuators.Actuator5 = 0.0;
+      jacoPoint.Position.Actuators.Actuator6 = 0.0;
+    }
   }
-  EraseAllTrajectories();
+
   //populate finger command
   if (msg.fingerCommand)
   {
@@ -960,6 +985,92 @@ void JacoArmTrajectoryController::cartesianCmdCallback(const wpi_jaco_msgs::Cart
         rate.sleep();
       }
     }
+  }
+}
+
+void JacoArmTrajectoryController::fingerPositionControl(float f1, float f2, float f3)
+{
+  f1 = max(f1, .02f);
+  f2 = max(f2, .02f);
+  f3 = max(f3, .02f);
+
+  TrajectoryPoint jacoPoint;
+  jacoPoint.InitStruct();
+  jacoPoint.Position.Type = ANGULAR_VELOCITY;
+  jacoPoint.Position.Actuators.Actuator1 = 0.0;
+  jacoPoint.Position.Actuators.Actuator2 = 0.0;
+  jacoPoint.Position.Actuators.Actuator3 = 0.0;
+  jacoPoint.Position.Actuators.Actuator4 = 0.0;
+  jacoPoint.Position.Actuators.Actuator5 = 0.0;
+  jacoPoint.Position.Actuators.Actuator6 = 0.0;
+  jacoPoint.Position.HandMode = VELOCITY_MODE;
+
+  bool goalReached = false;
+  AngularPosition position_data;
+  float error[3];
+  float prevTotalError;
+  float counter = 0; //check if error is unchanging, this likely means a finger is blocked by something so the controller should terminate
+  vector<float> errorFinger1;
+  vector<float> errorFinger2;
+  vector<float> errorFinger3;
+  errorFinger1.resize(10);
+  errorFinger2.resize(10);
+  errorFinger3.resize(10);
+  for (unsigned int i = 0; i < errorFinger1.size(); i ++)
+  {
+    errorFinger1[i] = 0.0;
+    errorFinger2[i] = 0.0;
+    errorFinger3[i] = 0.0;
+  }
+  ros::Rate rate(600);
+  while (!goalReached)
+  {
+    //get current finger position
+    GetAngularPosition(position_data);
+    error[0] = f1 - position_data.Fingers.Finger1;
+    error[1] = f2 - position_data.Fingers.Finger2;
+    error[2] = f3 - position_data.Fingers.Finger3;
+
+    float totalError = sqrt(pow(error[0], 2) + pow(error[1], 2) + pow(error[2], 2));
+    if (totalError == prevTotalError)
+    {
+      counter ++;
+    }
+    else
+    {
+      counter = 0;
+      prevTotalError = totalError;
+    }
+
+    if (totalError < FINGER_ERROR_THRESHOLD || counter > 40)
+    {
+      goalReached = true;
+      jacoPoint.Position.Fingers.Finger1 = 0.0;
+      jacoPoint.Position.Fingers.Finger2 = 0.0;
+      jacoPoint.Position.Fingers.Finger3 = 0.0;
+    }
+    else
+    {
+      float errorSum[3] = {0};
+      for (unsigned int i = 0; i < errorFinger1.size(); i ++)
+      {
+        errorSum[0] += errorFinger1[i];
+        errorSum[1] += errorFinger2[i];
+        errorSum[2] += errorFinger3[i];
+      }
+      jacoPoint.Position.Fingers.Finger1 = max(min(KP_F*error[0] + KV_F*(error[0] - errorFinger1.front()) + KI_F*errorSum[0], 30.0), -30.0);
+      jacoPoint.Position.Fingers.Finger2 = max(min(KP_F*error[1] + KV_F*(error[1] - errorFinger2.front()) + KI_F*errorSum[1], 30.0), -30.0);
+      jacoPoint.Position.Fingers.Finger3 = max(min(KP_F*error[2] + KV_F*(error[2] - errorFinger3.front()) + KI_F*errorSum[2], 30.0), -30.0);
+      errorFinger1.insert(errorFinger1.begin(), error[0]);
+      errorFinger2.insert(errorFinger2.begin(), error[1]);
+      errorFinger3.insert(errorFinger3.begin(), error[2]);
+      errorFinger1.resize(10);
+      errorFinger2.resize(10);
+      errorFinger3.resize(10);
+    }
+    EraseAllTrajectories();
+    SendBasicTrajectory(jacoPoint);
+    rate.sleep();
   }
 }